Version 1.0.
Fixed poses of the text.
Fixed bug in the method writing(): if there are not turtles alive the turtle doesn't try to follow anyone.
Updated state machine.
Fixed errors in get_closer_turtle() and become_angry(): now the dictionary is used correctly and the turtles are deleted fro the dictionary after they are dead.